feat: switched project file structure to more closely match next.js commonality
