+++ src/app/blog/layout.tsx
@@ -1,0 +1,7 @@
+export default function BlogLayout({
+    children,
+  }: {
+    children: React.ReactNode
+  }) {
+    return <section>{children}</section>
+  }
