+++ src/app/blog/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>Blog coming soon! Switched to use Next.js hehe, you get to hear me yap soon!</div>
+  )
+}
+
+export default page
