+++ src/components/Introduction/Introduction.tsx
@@ -6,7 +6,7 @@ import HomeWorkIcon from '@mui/icons-material/HomeWork';
 import LocationCityIcon from '@mui/icons-material/LocationCity';
 import Links from "../Links/Links"
 import PROFILEGIF from '../../../public/profileGif.gif'
-import { introInfo } from "../../services/constants";
+import { introInfo } from "../../lib/constants";
 import Image from 'next/image';
 import UoI from '../../../public/university-of-illinois.svg'
 const Introduction = () => {
