+++ src/lib/getStaticProps.ts
